create column django app
