+++ backend/account/query.py
@@ -5,7 +5,7 @@ Query functions for account related operations.
 """
 
 from django.core.exceptions import ObjectDoesNotExist
-from account.models import User, KanbanColumn
+from account.models import User
 
 
 def get_or_create_user(payload: dict) -> User:
@@ -17,7 +17,9 @@ def get_or_create_user(payload: dict) -> User:
             google_id=payload["sub"],
             email=payload["email"] if payload.get("email") else "No Email Found",
             image_url=payload["picture"] if payload.get("picture") else None,
-            first_name=payload["given_name"] if payload.get("given_name") else "No First Name Found",
+            first_name=payload["given_name"]
+            if payload.get("given_name")
+            else "No First Name Found",
             last_name=payload.get("family_name") if payload.get("family_name") else "",
         )
 
@@ -40,20 +42,3 @@ def update_user(payload: dict):
     # isn't raised. That is, we'll only save changes if the entire payload is
     # error-free.
     user.save()
-
-
-def create_column(user_id: int, column_name: str, column_number: int) -> KanbanColumn:
-    return KanbanColumn.objects.create(user=User.objects.get(id=user_id), name=column_name, column_number=column_number)
-
-
-def get_columns(user_id: int) -> list[KanbanColumn]:
-    return KanbanColumn.objects.filter(user=User.objects.get(id=user_id))
-
-
-def delete_column(user_id: int, column_number: int):
-    KanbanColumn.objects.get(user=user_id, column_number=column_number).delete()
-
-
-def delete_columns(ids: list[int]):
-    for column_id in ids:
-        KanbanColumn.objects.filter(id=column_id).delete()
