+++ backend/jamco/urls.py
@@ -19,6 +19,7 @@ from django.conf import settings
 from django.conf.urls.static import static
 
 from account import views as account_views
+from column import views as column_views
 from job import views as job_views
 from .views import index
 
@@ -36,13 +37,15 @@ urlpatterns = [
         account_views.update_account,
         name="update_account",
     ),
-    path("account/api/get_columns", account_views.get_columns, name="get_columns"),
+    path("column/api/get_columns", column_views.get_columns, name="get_columns"),
     path(
-        "account/api/update_columns",
-        account_views.update_columns,
+        "column/api/update_columns",
+        column_views.update_columns,
         name="update_columns",
     ),
-    path("job/api/get_minimum_jobs", job_views.get_minimum_jobs, name="get_minimum_jobs"),
+    path(
+        "job/api/get_minimum_jobs", job_views.get_minimum_jobs, name="get_minimum_jobs"
+    ),
     path("job/api/get_job_by_id", job_views.get_job_by_id, name="get_job_by_id"),
     path("job/api/create_job", job_views.create_job, name="create_job"),
     path("job/api/update_job", job_views.update_job, name="update_job"),
